+++ tests/python/test_structural_tag_converter.py
@@ -1,3 +1,4 @@
+import json
 import sys
 import time
 from typing import Any, Dict, List, Optional, Tuple, Union
@@ -67,10 +68,10 @@ def disable_profiler(request):
         profiler = Profiler(tokenizer_id)
 
 
-def check_stag_with_grammar(structural_tag_format: Dict[str, Any], expected_grammar_ebnf: str):
+def check_stag_with_grammar(structural_tag_format: Dict[str, Any], expected_stag_ebnf: str):
     structural_tag = {"type": "structural_tag", "format": structural_tag_format}
     stag_ebnf = xgr.Grammar.from_structural_tag(structural_tag)
-    assert str(stag_ebnf) == expected_grammar_ebnf
+    assert str(stag_ebnf) == expected_stag_ebnf
 
 
 def check_stag_with_instance(
@@ -90,41 +91,20 @@ def check_stag_with_instance(
         profiler.profile_stag(structural_tag_format, instance)
 
 
-def check_template_stag_with_grammar(
+def check_template_stag_with_stag(
     structural_tag_format: Union[Dict[str, Any], StructuralTag],
-    expected_grammar_ebnf: str,
+    expected_stag: Dict,
     **kwargs: List[Dict[str, str]],
 ):
     if isinstance(structural_tag_format, StructuralTag) or (
         "type" in structural_tag_format and structural_tag_format["type"] == "structural_tag"
     ):
-        grammar = xgr.Grammar.from_structural_tag_template(structural_tag_format, **kwargs)
+        stag = xgr.StructuralTag.from_template(structural_tag_format, **kwargs)
     else:
         structural_tag = {"type": "structural_tag", "format": structural_tag_format}
-        grammar = xgr.Grammar.from_structural_tag_template(structural_tag, **kwargs)
-    print(grammar)
-    print(expected_grammar_ebnf)
-    assert str(grammar) == expected_grammar_ebnf
-
-
-def check_template_stag_with_instance(
-    structural_tag_format: Union[Dict[str, Any], StructuralTag],
-    instance: str,
-    is_accepted: bool = True,
-    debug_print: bool = False,
-    **kwargs: List[Dict[str, str]],
-):
-    if isinstance(structural_tag_format, StructuralTag) or (
-        "type" in structural_tag_format and structural_tag_format["type"] == "structural_tag"
-    ):
-        stag_grammar = xgr.Grammar.from_structural_tag_template(structural_tag_format, **kwargs)
-    else:
-        structural_tag = {"type": "structural_tag", "format": structural_tag_format}
-        stag_grammar = xgr.Grammar.from_structural_tag_template(structural_tag, **kwargs)
-    accepted = _is_grammar_accept_string(stag_grammar, instance, debug_print=debug_print)
-    assert accepted == is_accepted
-    if PROFILER_ON:
-        profiler.profile_stag(structural_tag_format, instance)
+        stag = xgr.StructuralTag.from_template(structural_tag, **kwargs)
+    print(stag.model_dump())
+    assert json.dumps(expected_stag) == json.dumps(stag.model_dump())
 
 
 const_string_stag_grammar = [
@@ -144,12 +124,12 @@ const_string_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", const_string_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", const_string_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", const_string_instance_is_accepted)
 def test_const_string_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted, debug_print=True)
 
 
@@ -194,12 +174,12 @@ json_schema_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", json_schema_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", json_schema_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", json_schema_instance_is_accepted)
 def test_json_schema_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -266,12 +246,12 @@ qwen_parameter_xml_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", qwen_parameter_xml_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", qwen_parameter_xml_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", qwen_parameter_xml_instance_is_accepted)
 def test_qwen_parameter_xml_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -297,12 +277,12 @@ ebnf_grammar_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", ebnf_grammar_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", ebnf_grammar_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", ebnf_grammar_instance_is_accepted)
 def test_ebnf_grammar_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -324,12 +304,12 @@ regex_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", regex_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", regex_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", regex_instance_is_accepted)
 def test_regex_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -391,12 +371,12 @@ sequence_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", sequence_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", sequence_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", sequence_instance_is_accepted)
 def test_sequence_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -447,12 +427,12 @@ or_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", or_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", or_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", or_instance_is_accepted)
 def test_or_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -527,12 +507,12 @@ tag_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", tag_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", tag_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", tag_instance_is_accepted)
 def test_tag_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -560,12 +540,12 @@ any_text_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", any_text_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", any_text_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", any_text_instance_is_accepted)
 def test_any_text_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -582,12 +562,12 @@ root ::= ((any_text))
 any_text_only_instance_is_accepted = [("ABCDEF", True), ("123456", True), ("", True)]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", any_text_only_stag_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", any_text_only_stag_grammar)
 @pytest.mark.parametrize("instance, is_accepted", any_text_only_instance_is_accepted)
 def test_any_text_only_format(
-    stag_format: Dict[str, Any], expected_grammar: str, instance: str, is_accepted: bool
+    stag_format: Dict[str, Any], expected_stag: Dict, instance: str, is_accepted: bool
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -681,16 +661,16 @@ triggered_tag_instance_accepted_results = [
 ]
 
 
-@pytest.mark.parametrize("stag_id, stag_format, expected_grammar", triggered_tag_stag_grammar)
+@pytest.mark.parametrize("stag_id, stag_format, expected_stag", triggered_tag_stag_grammar)
 @pytest.mark.parametrize("instance, accepted_results", triggered_tag_instance_accepted_results)
 def test_triggered_tag_format(
     stag_id: int,
     stag_format: Dict[str, Any],
-    expected_grammar: str,
+    expected_stag: Dict,
     instance: str,
     accepted_results: List[bool],
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, accepted_results[stag_id])
 
 
@@ -724,15 +704,14 @@ root ::= ((triggered_tags))
 
 
 @pytest.mark.parametrize(
-    "stag_format, expected_grammar, instance_is_accepted_tuples",
-    test_triggered_tags_corner_case_data,
+    "stag_format, expected_stag, instance_is_accepted_tuples", test_triggered_tags_corner_case_data
 )
 def test_triggered_tags_corner_case(
     stag_format: Dict[str, Any],
-    expected_grammar: str,
+    expected_stag: Dict,
     instance_is_accepted_tuples: List[Tuple[str, bool]],
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     for instance, is_accepted in instance_is_accepted_tuples:
         check_stag_with_instance(stag_format, instance, is_accepted)
 
@@ -844,7 +823,7 @@ triggered_tag_with_outside_tag_instance_accepted_results = [
 
 
 @pytest.mark.parametrize(
-    "stag_id, stag_format, expected_grammar", triggered_tag_with_outside_tag_stag_grammar
+    "stag_id, stag_format, expected_stag", triggered_tag_with_outside_tag_stag_grammar
 )
 @pytest.mark.parametrize(
     "instance, accepted_results", triggered_tag_with_outside_tag_instance_accepted_results
@@ -852,11 +831,11 @@ triggered_tag_with_outside_tag_instance_accepted_results = [
 def test_triggered_tag_with_outside_tag(
     stag_id: int,
     stag_format: Dict[str, Any],
-    expected_grammar: str,
+    expected_stag: Dict,
     instance: str,
     accepted_results: List[bool],
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, accepted_results[stag_id])
 
 
@@ -935,18 +914,18 @@ tags_with_separator_instance_accepted_results = [
 ]
 
 
-@pytest.mark.parametrize("stag_id, stag_format, expected_grammar", tags_with_separator_stag_grammar)
+@pytest.mark.parametrize("stag_id, stag_format, expected_stag", tags_with_separator_stag_grammar)
 @pytest.mark.parametrize(
     "instance, accepted_results", tags_with_separator_instance_accepted_results
 )
 def test_tags_with_separator_format(
     stag_id: int,
     stag_format: Dict[str, Any],
-    expected_grammar: str,
+    expected_stag: Dict,
     instance: str,
     accepted_results: List[bool],
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, accepted_results[stag_id])
 
 
@@ -1038,7 +1017,7 @@ tags_with_separator_with_outside_tag_instance_accepted_results = [
 
 
 @pytest.mark.parametrize(
-    "stag_id, stag_format, expected_grammar", tags_with_separator_with_outside_tag_stag_grammar
+    "stag_id, stag_format, expected_stag", tags_with_separator_with_outside_tag_stag_grammar
 )
 @pytest.mark.parametrize(
     "instance, accepted_results", tags_with_separator_with_outside_tag_instance_accepted_results
@@ -1046,11 +1025,11 @@ tags_with_separator_with_outside_tag_instance_accepted_results = [
 def test_tags_with_separator_format_with_outside_tag(
     stag_id: int,
     stag_format: Dict[str, Any],
-    expected_grammar: str,
+    expected_stag: Dict,
     instance: str,
     accepted_results: List[bool],
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, accepted_results[stag_id])
 
 
@@ -1135,7 +1114,7 @@ tags_with_empty_separator_instance_accepted_results = [
 
 
 @pytest.mark.parametrize(
-    "stag_id, stag_format, expected_grammar", tags_with_empty_separator_stag_grammar
+    "stag_id, stag_format, expected_stag", tags_with_empty_separator_stag_grammar
 )
 @pytest.mark.parametrize(
     "instance, accepted_results", tags_with_empty_separator_instance_accepted_results
@@ -1143,11 +1122,11 @@ tags_with_empty_separator_instance_accepted_results = [
 def test_tags_with_empty_separator_format(
     stag_id: int,
     stag_format: Dict[str, Any],
-    expected_grammar: str,
+    expected_stag: Dict,
     instance: str,
     accepted_results: List[bool],
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, accepted_results[stag_id])
 
 
@@ -1590,14 +1569,14 @@ root ::= ((or_1))
 
 
 @pytest.mark.parametrize(
-    "stag_format, expected_grammar, instance_is_accepted_tuples", end_string_detector_test_data
+    "stag_format, expected_stag, instance_is_accepted_tuples", end_string_detector_test_data
 )
 def test_end_string_detector(
     stag_format: Dict[str, Any],
-    expected_grammar: str,
+    expected_stag: Dict,
     instance_is_accepted_tuples: List[Tuple[str, bool]],
 ):
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     for instance, is_accepted in instance_is_accepted_tuples:
         check_stag_with_instance(stag_format, instance, is_accepted)
 
@@ -2171,9 +2150,9 @@ multiple_end_tokens_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", multiple_end_tokens_tag_stag_grammar)
-def test_multiple_end_tokens_tag_grammar(stag_format: Dict[str, Any], expected_grammar: str):
-    check_stag_with_grammar(stag_format, expected_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", multiple_end_tokens_tag_stag_grammar)
+def test_multiple_end_tokens_tag_grammar(stag_format: Dict[str, Any], expected_stag: Dict):
+    check_stag_with_grammar(stag_format, expected_stag)
 
 
 @pytest.mark.parametrize("instance, is_accepted", multiple_end_tokens_instance_is_accepted)
@@ -2216,9 +2195,9 @@ multiple_end_tokens_any_text_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize("stag_format, expected_grammar", multiple_end_tokens_any_text_stag_grammar)
-def test_multiple_end_tokens_any_text_grammar(stag_format: Dict[str, Any], expected_grammar: str):
-    check_stag_with_grammar(stag_format, expected_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", multiple_end_tokens_any_text_stag_grammar)
+def test_multiple_end_tokens_any_text_grammar(stag_format: Dict[str, Any], expected_stag: Dict):
+    check_stag_with_grammar(stag_format, expected_stag)
 
 
 @pytest.mark.parametrize("instance, is_accepted", multiple_end_tokens_any_text_instance_is_accepted)
@@ -2273,11 +2252,9 @@ multiple_end_tokens_with_empty_instance_is_accepted = [
 ]
 
 
-@pytest.mark.parametrize(
-    "stag_format, expected_grammar", multiple_end_tokens_with_empty_stag_grammar
-)
-def test_multiple_end_tokens_with_empty_grammar(stag_format: Dict[str, Any], expected_grammar: str):
-    check_stag_with_grammar(stag_format, expected_grammar)
+@pytest.mark.parametrize("stag_format, expected_stag", multiple_end_tokens_with_empty_stag_grammar)
+def test_multiple_end_tokens_with_empty_grammar(stag_format: Dict[str, Any], expected_stag: Dict):
+    check_stag_with_grammar(stag_format, expected_stag)
 
 
 @pytest.mark.parametrize(
@@ -2374,7 +2351,7 @@ def test_excluded_strings_in_any_text(instance: str, is_accepted: bool):
         "end": ".",
     }
 
-    expected_grammar = r"""any_text ::= TagDispatch(
+    expected_stag = r"""any_text ::= TagDispatch(
   stop_eos=false,
   stop_str=("."),
   loop_after_dispatch=false,
@@ -2384,7 +2361,7 @@ tag ::= (("" any_text))
 root ::= ((tag))
 """
 
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -2415,7 +2392,7 @@ def test_excluded_strings_in_triggered_format(instance: str, is_accepted: bool):
         "excludes": ["L1", "L2"],
     }
 
-    expected_grammar = r"""const_string ::= (("L1"))
+    expected_stag = r"""const_string ::= (("L1"))
 const_string_1 ::= (("L2"))
 triggered_tags_group ::= (("1" const_string "A") | ("2" const_string_1 "A"))
 triggered_tags_first ::= (("A1" const_string "A") | ("A2" const_string_1 "A"))
@@ -2430,7 +2407,7 @@ triggered_tags ::= ((triggered_tags_first triggered_tags_sub))
 root ::= ((triggered_tags))
 """
 
-    check_stag_with_grammar(stag_format, expected_grammar)
+    check_stag_with_grammar(stag_format, expected_stag)
     check_stag_with_instance(stag_format, instance, is_accepted)
 
 
@@ -2438,94 +2415,89 @@ const_string_template_values_stag_grammar_instance_accepted = [
     (
         {"type": "const_string", "value": "The value is: {{strings[].value}}."},
         [{"value": "a"}, {"value": "b"}, {"value": "c"}],
-        r"""const_string ::= (("The value is: a."))
-const_string_1 ::= (("The value is: b."))
-const_string_2 ::= (("The value is: c."))
-or ::= ((const_string) | (const_string_1) | (const_string_2))
-root ::= ((or))
-""",
-        [
-            ("The value is: a.", True),
-            ("The value is: b.", True),
-            ("The value is: c.", True),
-            ("The value is: d.", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "const_string", "value": "The value is: a."},
+                    {"type": "const_string", "value": "The value is: b."},
+                    {"type": "const_string", "value": "The value is: c."},
+                ],
+            },
+        },
     ),
     (
         {"type": "const_string", "value": "{{strings[].value}}"},
         [{"value": "a"}, {"value": "b"}, {"value": "c"}],
-        r"""const_string ::= (("a"))
-const_string_1 ::= (("b"))
-const_string_2 ::= (("c"))
-or ::= ((const_string) | (const_string_1) | (const_string_2))
-root ::= ((or))
-""",
-        [("a", True), ("b", True), ("c", True), ("d", False)],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "const_string", "value": "a"},
+                    {"type": "const_string", "value": "b"},
+                    {"type": "const_string", "value": "c"},
+                ],
+            },
+        },
     ),
     (
         {"type": "const_string", "value": "The value is: {{strings[].value}}"},
         [{"value": "a"}, {"value": "b"}, {"value": "c"}],
-        r"""const_string ::= (("The value is: a"))
-const_string_1 ::= (("The value is: b"))
-const_string_2 ::= (("The value is: c"))
-or ::= ((const_string) | (const_string_1) | (const_string_2))
-root ::= ((or))
-""",
-        [
-            ("The value is: a", True),
-            ("The value is: b", True),
-            ("The value is: c", True),
-            ("The value is: d", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "const_string", "value": "The value is: a"},
+                    {"type": "const_string", "value": "The value is: b"},
+                    {"type": "const_string", "value": "The value is: c"},
+                ],
+            },
+        },
     ),
     (
         {"type": "const_string", "value": "{{strings[].value}}是"},
         [{"value": "a"}, {"value": "b"}, {"value": "c"}],
-        r"""const_string ::= (("a\u662f"))
-const_string_1 ::= (("b\u662f"))
-const_string_2 ::= (("c\u662f"))
-or ::= ((const_string) | (const_string_1) | (const_string_2))
-root ::= ((or))
-""",
-        [("a是", True), ("b是", True), ("c是", True), ("d是", False)],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "const_string", "value": "a是"},
+                    {"type": "const_string", "value": "b是"},
+                    {"type": "const_string", "value": "c是"},
+                ],
+            },
+        },
     ),
     (
         {"type": "const_string", "value": "{{strings[].begin}} a dog{{strings[].end}}"},
         [{"begin": "It is", "end": "."}, {"begin": "Is it", "end": "?"}],
-        r"""const_string ::= (("It is a dog."))
-const_string_1 ::= (("Is it a dog\?"))
-or ::= ((const_string) | (const_string_1))
-root ::= ((or))
-""",
-        [
-            ("It is a dog.", True),
-            ("Is it a dog?", True),
-            ("It is a dog?", False),
-            ("Is it a dog.", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "const_string", "value": "It is a dog."},
+                    {"type": "const_string", "value": "Is it a dog?"},
+                ],
+            },
+        },
     ),
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     const_string_template_values_stag_grammar_instance_accepted,
 )
 def test_const_string_template_values(
-    template_stag_format: Dict[str, Any],
-    template_values: List[Dict[str, Any]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    template_stag_format: Dict[str, Any], template_values: List[Dict[str, Any]], expected_stag: Dict
 ):
     """Test const_string format with template values"""
-    check_template_stag_with_grammar(
-        template_stag_format, expected_grammar, strings=template_values
-    )
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, strings=template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, strings=template_values)
 
 
 def test_const_string_template_values_with_mingled_templates():
@@ -2555,82 +2527,36 @@ json_schema_template_values_stag_grammar_instance_accepted = [
             },
             {"value": r"""{"type":"string"}"""},
         ],
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_0 ::= (("{" [ \n\t]* "\"arg\"" [ \n\t]* ":" [ \n\t]* basic_string [ \n\t]* "}"))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= ((basic_string_1))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-or ::= ((root_0) | (root_1))
-root ::= ((or))
-""",
-        [
-            ('{"arg": "value"}', True),
-            ('{"arg": "another value"}', True),
-            ('{"arg": 123}', False),
-            ('{"arg": "value", "extra": "field"}', False),
-            ('{"arg": "value"', False),
-            ('"just a string"', True),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {
+                        "type": "json_schema",
+                        "json_schema": {
+                            "type": "object",
+                            "properties": {"arg": {"type": "string"}},
+                            "required": ["arg"],
+                        },
+                    },
+                    {"type": "json_schema", "json_schema": {"type": "string"}},
+                ],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     json_schema_template_values_stag_grammar_instance_accepted,
 )
 def test_json_schema_template_values(
-    template_stag_format: Dict[str, Any],
-    template_values: List[Dict[str, Any]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    template_stag_format: Dict[str, Any], template_values: List[Dict[str, Any]], expected_stag: Dict
 ):
     """Test json_schema format with template values"""
-    check_template_stag_with_grammar(
-        template_stag_format, expected_grammar, schemas=template_values
-    )
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, schemas=template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, schemas=template_values)
 
 
 def test_part_json_schema_template_failure():
@@ -2657,104 +2583,43 @@ qwen_template_values_stag_grammar_instance_accepted = [
                 "value": r"""{"type":"object", "properties": {"age": {"type": "integer"}}, "required": ["age"]}"""
             },
         ],
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-xml_string ::= TagDispatch(
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=false,
-  excludes=("</parameter>")
-)
-xml_variable_name ::= (([a-zA-Z_] [a-zA-Z0-9_]*))
-xml_string_0 ::= ((xml_string))
-xml_any ::= ((basic_number) | (xml_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_0 ::= (([ \n\t]* "<parameter=name>" [ \n\t]* xml_string_0 [ \n\t]* "</parameter>"))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-xml_string_1 ::= TagDispatch(
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=false,
-  excludes=("</parameter>")
-)
-xml_variable_name_1 ::= (([a-zA-Z_] [a-zA-Z0-9_]*))
-xml_string_0_1 ::= ((xml_string_1))
-xml_any_1 ::= ((basic_number_8) | (xml_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_prop_0 ::= (("0") | (root_prop_0_1 [1-9] [0-9]*))
-root_1 ::= (([ \n\t]* "<parameter=age>" [ \n\t]* root_prop_0 [ \n\t]* "</parameter>"))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-root_prop_0_1 ::= ("" | ("-"))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-or ::= ((root_0) | (root_1))
-root ::= ((or))
-""",
-        [
-            ("<parameter=name>value</parameter>", True),
-            ("<parameter=name>another value</parameter>", True),
-            ("<parameter=name>123</parameter>", True),
-            ("<parameter=name>value</parameter>", True),
-            ("just a string", False),
-            ("<parameter=age>25</parameter>", True),
-            ("<parameter=age>-5</parameter>", True),
-            ("<parameter=age>abc</parameter>", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {
+                        "type": "qwen_xml_parameter",
+                        "json_schema": {
+                            "type": "object",
+                            "properties": {"name": {"type": "string"}},
+                            "required": ["name"],
+                        },
+                    },
+                    {
+                        "type": "qwen_xml_parameter",
+                        "json_schema": {
+                            "type": "object",
+                            "properties": {"age": {"type": "integer"}},
+                            "required": ["age"],
+                        },
+                    },
+                ],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     qwen_template_values_stag_grammar_instance_accepted,
 )
 def test_qwen_template_values(
-    template_stag_format: Dict[str, Any],
-    template_values: List[Dict[str, Any]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    template_stag_format: Dict[str, Any], template_values: List[Dict[str, Any]], expected_stag: Dict
 ):
     """Test qwen_xml_parameter format with template values"""
-    check_template_stag_with_grammar(
-        template_stag_format, expected_grammar, schemas=template_values
-    )
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, schemas=template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, schemas=template_values)
 
 
 def test_part_qwen_template_failure():
@@ -2778,36 +2643,29 @@ regex_template_values_stag_grammar_instance_accepted = [
     (
         {"type": "regex", "pattern": r"{{patterns[].value}}"},
         [{"value": r"123"}, {"value": r"[a-zA-Z]+"}],
-        r"""root_0 ::= (("1" "2" "3"))
-root_1 ::= ((root_1_1))
-root_1_1 ::= (([a-zA-Z] root_1_1) | ([a-zA-Z]))
-or ::= ((root_0) | (root_1))
-root ::= ((or))
-""",
-        [("123", True), ("abc", True), ("123abc", False), ("123abc456", False), ("", False)],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "regex", "pattern": "123"},
+                    {"type": "regex", "pattern": "[a-zA-Z]+"},
+                ],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     regex_template_values_stag_grammar_instance_accepted,
 )
 def test_regex_template_values(
-    template_stag_format: Dict[str, Any],
-    template_values: List[Dict[str, Any]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    template_stag_format: Dict[str, Any], template_values: List[Dict[str, Any]], expected_stag: Dict
 ):
     """Test regex format with template values"""
-    check_template_stag_with_grammar(
-        template_stag_format, expected_grammar, patterns=template_values
-    )
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, patterns=template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, patterns=template_values)
 
 
 def test_part_regex_template_failure():
@@ -2823,45 +2681,29 @@ grammar_template_values_stag_grammar_instance_accepted = [
     (
         {"type": "grammar", "grammar": "{{grammars[].value}}"},
         [{"value": 'root::= "a" | "b"'}, {"value": 'root ::= a+\na::= "c" | "d"'}],
-        r"""root_0 ::= (("a") | ("b"))
-root_1 ::= ((root_1_1))
-a ::= (("c") | ("d"))
-root_1_1 ::= ((a root_1_1) | (a))
-or ::= ((root_0) | (root_1))
-root ::= ((or))
-""",
-        [
-            ("a", True),
-            ("b", True),
-            ("c", True),
-            ("d", True),
-            ("aa", False),
-            ("ab", False),
-            ("cc", True),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "grammar", "grammar": 'root::= "a" | "b"'},
+                    {"type": "grammar", "grammar": 'root ::= a+\na::= "c" | "d"'},
+                ],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     grammar_template_values_stag_grammar_instance_accepted,
 )
 def test_grammar_template_values(
-    template_stag_format: Dict[str, Any],
-    template_values: List[Dict[str, Any]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    template_stag_format: Dict[str, Any], template_values: List[Dict[str, Any]], expected_stag: Dict
 ):
     """Test grammar format with template values"""
-    check_template_stag_with_grammar(
-        template_stag_format, expected_grammar, grammars=template_values
-    )
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, grammars=template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, grammars=template_values)
 
 
 def test_part_grammar_template_failure():
@@ -2874,23 +2716,6 @@ def test_part_grammar_template_failure():
 
 
 any_text_instance_is_accepted_template = [("abc", True), ("好", True)]
-
-
-@pytest.mark.parametrize("instance, is_accepted", any_text_instance_is_accepted_template)
-def test_any_text_compatible(instance: str, is_accepted: bool):
-    """Test that AnyTextFormat is compatible with all structural tag formats"""
-    any_text_format = {"type": "any_text"}
-    dummy_grammars = [{"value": 'root ::= "a" | "b"'}, {"value": 'root ::= "c" | "d"'}]
-
-    expected_grammar = r"""any_text ::= (([\0-\U0010ffff]*))
-root ::= ((any_text))
-"""
-
-    check_template_stag_with_grammar(any_text_format, expected_grammar, grammars=dummy_grammars)
-
-    check_template_stag_with_instance(
-        any_text_format, instance, is_accepted, grammars=dummy_grammars
-    )
 
 
 def test_no_parameter_error():
@@ -2917,42 +2742,33 @@ or_template_values_stag_grammar_instance_accepted = [
             "strings": [{"value": "hello"}, {"value": "world"}],
             "numbers": [{"value": "1"}, {"value": "2"}],
         },
-        r"""const_string ::= (("hello"))
-const_string_1 ::= (("world"))
-const_string_2 ::= (("1"))
-const_string_3 ::= (("2"))
-or ::= ((const_string) | (const_string_1) | (const_string_2) | (const_string_3))
-root ::= ((or))
-""",
-        [
-            ("hello", True),
-            ("world", True),
-            ("1", True),
-            ("2", True),
-            ("3", False),
-            ("hello world", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {"type": "const_string", "value": "hello"},
+                    {"type": "const_string", "value": "world"},
+                    {"type": "const_string", "value": "1"},
+                    {"type": "const_string", "value": "2"},
+                ],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     or_template_values_stag_grammar_instance_accepted,
 )
 def test_or_template_values(
     template_stag_format: Dict[str, Any],
     template_values: Dict[str, List[Dict[str, Any]]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    expected_stag: Dict,
 ):
     """Test grammar format with template values"""
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, **template_values)
 
 
 sequence_template_values_stag_grammar_instance_accepted = [
@@ -2968,44 +2784,43 @@ sequence_template_values_stag_grammar_instance_accepted = [
             "first": [{"value": "I'm "}, {"value": "You're "}],
             "second": [{"value": "Alice"}, {"value": "Bob"}],
         },
-        r"""const_string ::= (("I\'m "))
-const_string_1 ::= (("You\'re "))
-or ::= ((const_string) | (const_string_1))
-const_string_2 ::= (("Alice"))
-const_string_3 ::= (("Bob"))
-or_1 ::= ((const_string_2) | (const_string_3))
-sequence ::= ((or or_1))
-root ::= ((sequence))
-""",
-        [
-            ("I'm Alice", True),
-            ("You're Bob", True),
-            ("I'm Bob", True),
-            ("You're Alice", True),
-            ("Alice I'm", False),
-            ("Bob You're", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "sequence",
+                "elements": [
+                    {
+                        "type": "or",
+                        "elements": [
+                            {"type": "const_string", "value": "I'm "},
+                            {"type": "const_string", "value": "You're "},
+                        ],
+                    },
+                    {
+                        "type": "or",
+                        "elements": [
+                            {"type": "const_string", "value": "Alice"},
+                            {"type": "const_string", "value": "Bob"},
+                        ],
+                    },
+                ],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     sequence_template_values_stag_grammar_instance_accepted,
 )
 def test_sequence_template_values(
     template_stag_format: Dict[str, Any],
     template_values: Dict[str, List[Dict[str, Any]]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    expected_stag: Dict,
 ):
     """Test sequence format with template values"""
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, **template_values)
 
 
 tag_template_values_stag_grammar_instance_accepted = [
@@ -3023,48 +2838,53 @@ tag_template_values_stag_grammar_instance_accepted = [
             "end": "{{outter[].symbol}}",
         },
         {"outter": [{"first": "It is a ", "symbol": "!"}, {"first": "Is it a ", "symbol": "?"}]},
-        r"""const_string ::= (("dog"))
-const_string_1 ::= (("cat"))
-or ::= ((const_string) | (const_string_1))
-tag ::= (("It is a " or "!"))
-const_string_2 ::= (("dog"))
-const_string_3 ::= (("cat"))
-or_1 ::= ((const_string_2) | (const_string_3))
-tag_1 ::= (("Is it a " or_1 "\?"))
-or_2 ::= ((tag) | (tag_1))
-root ::= ((or_2))
-""",
-        [
-            ("It is a dog!", True),
-            ("Is it a cat?", True),
-            ("It is a cat!", True),
-            ("Is it a dog?", True),
-            ("It is a dog?", False),
-            ("It is a cat?", False),
-            ("Is it a dog!", False),
-            ("Is it a cat!", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "or",
+                "elements": [
+                    {
+                        "type": "tag",
+                        "begin": "It is a ",
+                        "content": {
+                            "type": "or",
+                            "elements": [
+                                {"type": "const_string", "value": "dog"},
+                                {"type": "const_string", "value": "cat"},
+                            ],
+                        },
+                        "end": ["!"],
+                    },
+                    {
+                        "type": "tag",
+                        "begin": "Is it a ",
+                        "content": {
+                            "type": "or",
+                            "elements": [
+                                {"type": "const_string", "value": "dog"},
+                                {"type": "const_string", "value": "cat"},
+                            ],
+                        },
+                        "end": ["?"],
+                    },
+                ],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     tag_template_values_stag_grammar_instance_accepted,
 )
 def test_tag_template_values(
     template_stag_format: Dict[str, Any],
     template_values: Dict[str, List[Dict[str, Any]]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    expected_stag: Dict,
 ):
     """Test tag format with template values"""
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, **template_values)
 
 
 def test_mingled_tag_template():
@@ -3109,55 +2929,57 @@ triggered_tag_template_values_stag_grammar_instance_accepted = [
             "stop_after_first": False,
         },
         {"outter": [{"first": "It is a ", "symbol": "!"}, {"first": "Is it a ", "symbol": "?"}]},
-        r"""const_string ::= (("dog"))
-const_string_1 ::= (("cat"))
-or ::= ((const_string) | (const_string_1))
-const_string_2 ::= (("dog"))
-const_string_3 ::= (("cat"))
-or_1 ::= ((const_string_2) | (const_string_3))
-triggered_tags_group ::= (("t is a " or "!") | ("s it a " or_1 "\?"))
-triggered_tags ::= TagDispatch(
-  ("I", triggered_tags_group),
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=true,
-  excludes=()
-)
-root ::= ((triggered_tags))
-""",
-        [
-            ("It is a dog!", True),
-            ("Is it a cat?", True),
-            ("It is a cat!", True),
-            ("Is it a dog?", True),
-            ("It is a dog?", False),
-            ("It is a cat?", False),
-            ("Is it a dog!", False),
-            ("Is it a cat!", False),
-            ("Hello world", True),
-            ("I am happy", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "triggered_tags",
+                "triggers": ["I"],
+                "tags": [
+                    {
+                        "type": "tag",
+                        "begin": "It is a ",
+                        "content": {
+                            "type": "or",
+                            "elements": [
+                                {"type": "const_string", "value": "dog"},
+                                {"type": "const_string", "value": "cat"},
+                            ],
+                        },
+                        "end": ["!"],
+                    },
+                    {
+                        "type": "tag",
+                        "begin": "Is it a ",
+                        "content": {
+                            "type": "or",
+                            "elements": [
+                                {"type": "const_string", "value": "dog"},
+                                {"type": "const_string", "value": "cat"},
+                            ],
+                        },
+                        "end": ["?"],
+                    },
+                ],
+                "at_least_one": False,
+                "stop_after_first": False,
+                "excludes": [],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     triggered_tag_template_values_stag_grammar_instance_accepted,
 )
 def test_triggered_tag_template_values(
     template_stag_format: Dict[str, Any],
     template_values: Dict[str, List[Dict[str, Any]]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    expected_stag: Dict,
 ):
     """Test triggered_tags format with template values"""
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, **template_values)
 
 
 tag_with_separator_template_values_stag_grammar_instance_accepted = [
@@ -3183,863 +3005,56 @@ tag_with_separator_template_values_stag_grammar_instance_accepted = [
             "stop_after_first": False,
         },
         {"outter": [{"first": "It is a ", "symbol": "!"}, {"first": "Is it a ", "symbol": "?"}]},
-        r"""const_string ::= (("dog"))
-const_string_1 ::= (("cat"))
-or ::= ((const_string) | (const_string_1))
-tag ::= (("It is a " or "!"))
-const_string_2 ::= (("dog"))
-const_string_3 ::= (("cat"))
-or_1 ::= ((const_string_2) | (const_string_3))
-tag_1 ::= (("Is it a " or_1 "\?"))
-tags_with_separator_tags ::= ((tag) | (tag_1))
-tags_with_separator_sub ::= ("" | ("\n" tags_with_separator_tags tags_with_separator_sub))
-tags_with_separator ::= ("" | (tags_with_separator_tags tags_with_separator_sub))
-root ::= ((tags_with_separator))
-""",
-        [
-            ("It is a dog!", True),
-            ("Is it a cat?", True),
-            ("It is a cat!", True),
-            ("Is it a dog?", True),
-            ("It is a dog!\nIs it a cat?", True),
-            ("It is a cat!\nIt is a dog!", True),
-            ("Is it a dog?\nIt is a cat!", True),
-            ("Is it a cat?\nIs it a dog?\nIs it a cat?", True),
-            ("It is a dog?", False),
-            ("It is a cat!\nIt is a dog?", False),
-            ("Is it a dog!", False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "tags_with_separator",
+                "tags": [
+                    {
+                        "type": "tag",
+                        "begin": "It is a ",
+                        "content": {
+                            "type": "or",
+                            "elements": [
+                                {"type": "const_string", "value": "dog"},
+                                {"type": "const_string", "value": "cat"},
+                            ],
+                        },
+                        "end": ["!"],
+                    },
+                    {
+                        "type": "tag",
+                        "begin": "Is it a ",
+                        "content": {
+                            "type": "or",
+                            "elements": [
+                                {"type": "const_string", "value": "dog"},
+                                {"type": "const_string", "value": "cat"},
+                            ],
+                        },
+                        "end": ["?"],
+                    },
+                ],
+                "separator": "\n",
+                "at_least_one": False,
+                "stop_after_first": False,
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     tag_with_separator_template_values_stag_grammar_instance_accepted,
 )
 def test_tag_with_separator_template_values(
     template_stag_format: Dict[str, Any],
     template_values: Dict[str, List[Dict[str, Any]]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    expected_stag: Dict,
 ):
     """Test tags_with_separator format with template values"""
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
-
-
-builtin_template_values_stag_grammar_instance_accepted = [
-    (
-        "llama",
-        {
-            "tools": [
-                {
-                    "name": "Calculator",
-                    "description": "A calculator that can perform basic arithmetic operations.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "operation": {
-                                "type": "string",
-                                "enum": ["add", "subtract", "multiply", "divide"],
-                            },
-                            "a": {"type": "number"},
-                            "b": {"type": "number"},
-                        },
-                        "required": ["operation", "a", "b"],
-                    },
-                },
-                {
-                    "name": "Weather",
-                    "description": "A tool to get the current weather in a specified location.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "location": {
-                                "type": "string",
-                                "description": "The name of the city or location to get the weather for.",
-                            }
-                        },
-                        "required": ["location"],
-                    },
-                },
-            ]
-        },
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_prop_0 ::= (("\"add\"") | ("\"subtract\"") | ("\"multiply\"") | ("\"divide\""))
-root_part_1 ::= (([ \n\t]* "," [ \n\t]* "\"b\"" [ \n\t]* ":" [ \n\t]* basic_number))
-root_part_0 ::= (([ \n\t]* "," [ \n\t]* "\"a\"" [ \n\t]* ":" [ \n\t]* basic_number root_part_1))
-root_0 ::= (("{" [ \n\t]* "\"operation\"" [ \n\t]* ":" [ \n\t]* root_prop_0 root_part_0 [ \n\t]* "}"))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= (("{" [ \n\t]* "\"location\"" [ \n\t]* ":" [ \n\t]* basic_string_1 [ \n\t]* "}"))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-triggered_tags_group ::= (("\"Calculator\", \"parameters\": " root_0 "}") | ("\"Weather\", \"parameters\": " root_1 "}"))
-triggered_tags ::= TagDispatch(
-  ("{\"name\": ", triggered_tags_group),
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=true,
-  excludes=()
-)
-root ::= ((triggered_tags))
-""",
-        [
-            (
-                'OK, I will use the Calculator tool to perform the operation. {"name": "Calculator", "parameters": {"operation": "add", "a": 5, "b": 3}}',
-                True,
-            ),
-            (
-                'I need to know the weather in Paris. {"name": "Weather", "parameters": {"location": "Paris"}}',
-                True,
-            ),
-            (
-                'Can you calculate Paris add 5? {"name": "Calculator", "parameters": {"operation": "add", "a": 5, "b": "Paris"}}',
-                False,
-            ),
-            (
-                'I want to know the weather in 1. {"name": "Weather", "parameters": {"location": 1}}',
-                False,
-            ),
-            ("Some random text", True),
-        ],
-    ),
-    (
-        "kimi",
-        {
-            "tools": [
-                {
-                    "name": "Calculator",
-                    "description": "A calculator that can perform basic arithmetic operations.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "operation": {
-                                "type": "string",
-                                "enum": ["add", "subtract", "multiply", "divide"],
-                            },
-                            "a": {"type": "number"},
-                            "b": {"type": "number"},
-                        },
-                        "required": ["operation", "a", "b"],
-                    },
-                },
-                {
-                    "name": "Weather",
-                    "description": "A tool to get the current weather in a specified location.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "location": {
-                                "type": "string",
-                                "description": "The name of the city or location to get the weather for.",
-                            }
-                        },
-                        "required": ["location"],
-                    },
-                },
-            ]
-        },
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_prop_0 ::= (("\"add\"") | ("\"subtract\"") | ("\"multiply\"") | ("\"divide\""))
-root_part_1 ::= (([ \n\t]* "," [ \n\t]* "\"b\"" [ \n\t]* ":" [ \n\t]* basic_number))
-root_part_0 ::= (([ \n\t]* "," [ \n\t]* "\"a\"" [ \n\t]* ":" [ \n\t]* basic_number root_part_1))
-root_0 ::= (("{" [ \n\t]* "\"operation\"" [ \n\t]* ":" [ \n\t]* root_prop_0 root_part_0 [ \n\t]* "}"))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= (("{" [ \n\t]* "\"location\"" [ \n\t]* ":" [ \n\t]* basic_string_1 [ \n\t]* "}"))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-triggered_tags_group ::= (("Calculator<|tool_call_argument_begin|>" root_0 "<|tool_call_end|>") | ("Weather<|tool_call_argument_begin|>" root_1 "<|tool_call_end|>"))
-triggered_tags ::= TagDispatch(
-  ("<|tool_call_begin|>", triggered_tags_group),
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=true,
-  excludes=()
-)
-root ::= ((triggered_tags))
-""",
-        [
-            (
-                'OK, I will use the Calculator tool to perform the operation. <|tool_call_begin|>Calculator<|tool_call_argument_begin|>{"operation": "add", "a": 5, "b": 3}<|tool_call_end|>',
-                True,
-            ),
-            (
-                'I need to know the weather in Paris. <|tool_call_begin|>Weather<|tool_call_argument_begin|>{"location": "Paris"}<|tool_call_end|>',
-                True,
-            ),
-            (
-                'Can you calculate Paris add 5? <|tool_call_begin|>Calculator<|tool_call_argument_begin|>{"operation": "add", "a": 5, "b": "Paris"}<|tool_call_end|>',
-                False,
-            ),
-            (
-                'I want to know the weather in 1. <|tool_call_begin|>Weather<|tool_call_argument_begin|>{"location": 1}<|tool_call_end|>',
-                False,
-            ),
-            ("Some random text", True),
-        ],
-    ),
-    (
-        "deepseek",
-        {
-            "tools": [
-                {
-                    "name": "Calculator",
-                    "description": "A calculator that can perform basic arithmetic operations.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "operation": {
-                                "type": "string",
-                                "enum": ["add", "subtract", "multiply", "divide"],
-                            },
-                            "a": {"type": "number"},
-                            "b": {"type": "number"},
-                        },
-                        "required": ["operation", "a", "b"],
-                    },
-                },
-                {
-                    "name": "Weather",
-                    "description": "A tool to get the current weather in a specified location.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "location": {
-                                "type": "string",
-                                "description": "The name of the city or location to get the weather for.",
-                            }
-                        },
-                        "required": ["location"],
-                    },
-                },
-            ]
-        },
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_prop_0 ::= (("\"add\"") | ("\"subtract\"") | ("\"multiply\"") | ("\"divide\""))
-root_part_1 ::= (([ \n\t]* "," [ \n\t]* "\"b\"" [ \n\t]* ":" [ \n\t]* basic_number))
-root_part_0 ::= (([ \n\t]* "," [ \n\t]* "\"a\"" [ \n\t]* ":" [ \n\t]* basic_number root_part_1))
-root_0 ::= (("{" [ \n\t]* "\"operation\"" [ \n\t]* ":" [ \n\t]* root_prop_0 root_part_0 [ \n\t]* "}"))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= (("{" [ \n\t]* "\"location\"" [ \n\t]* ":" [ \n\t]* basic_string_1 [ \n\t]* "}"))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-triggered_tags_group ::= (("Calculator<\uff5ctool\u2581sep\uff5c>" root_0 "<\uff5ctool\u2581call\u2581end\uff5c>") | ("Weather<\uff5ctool\u2581sep\uff5c>" root_1 "<\uff5ctool\u2581call\u2581end\uff5c>"))
-triggered_tags ::= TagDispatch(
-  ("<\uff5ctool\u2581calls\u2581begin\uff5c><\uff5ctool\u2581call\u2581begin\uff5c>", triggered_tags_group),
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=true,
-  excludes=()
-)
-root ::= ((triggered_tags))
-""",
-        [
-            (
-                'OK, I will use the Calculator tool to perform the operation. <｜tool▁calls▁begin｜><｜tool▁call▁begin｜>Calculator<｜tool▁sep｜>{"operation": "add", "a": 5, "b": 3}<｜tool▁call▁end｜>',
-                True,
-            ),
-            (
-                'I need to know the weather in Paris. <｜tool▁calls▁begin｜><｜tool▁call▁begin｜>Weather<｜tool▁sep｜>{"location": "Paris"}<｜tool▁call▁end｜>',
-                True,
-            ),
-            (
-                'Can you calculate Paris add 5? <｜tool▁calls▁begin｜><｜tool▁call▁begin｜>Calculator<｜tool▁sep｜>{"operation": "add", "a": 5, "b": "Paris"}<｜tool▁call▁end｜>',
-                False,
-            ),
-            (
-                'I want to know the weather in 1. <｜tool▁calls▁begin｜><｜tool▁call▁begin｜>Weather<｜tool▁sep｜>{"location": 1}<<｜tool▁call▁end｜>',
-                False,
-            ),
-            ("Some random text", True),
-        ],
-    ),
-    (
-        "qwen_coder",
-        {
-            "tools": [
-                {
-                    "name": "Calculator",
-                    "description": "A calculator that can perform basic arithmetic operations.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "operation": {
-                                "type": "string",
-                                "enum": ["add", "subtract", "multiply", "divide"],
-                            },
-                            "a": {"type": "number"},
-                            "b": {"type": "number"},
-                        },
-                        "required": ["operation", "a", "b"],
-                    },
-                },
-                {
-                    "name": "Weather",
-                    "description": "A tool to get the current weather in a specified location.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "location": {
-                                "type": "string",
-                                "description": "The name of the city or location to get the weather for.",
-                            }
-                        },
-                        "required": ["location"],
-                    },
-                },
-            ]
-        },
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-xml_string ::= TagDispatch(
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=false,
-  excludes=("</parameter>")
-)
-xml_variable_name ::= (([a-zA-Z_] [a-zA-Z0-9_]*))
-xml_string_0 ::= ((xml_string))
-xml_any ::= ((basic_number) | (xml_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_prop_0 ::= (("add") | ("subtract") | ("multiply") | ("divide"))
-root_prop_1 ::= ((root_prop_1_1 root_prop_1_7 root_prop_1_3 root_prop_1_6))
-root_prop_2 ::= ((root_prop_2_1 root_prop_2_7 root_prop_2_3 root_prop_2_6))
-root_part_1 ::= (([ \n\t]* "<parameter=b>" [ \n\t]* root_prop_2 [ \n\t]* "</parameter>"))
-root_part_0 ::= (([ \n\t]* "<parameter=a>" [ \n\t]* root_prop_1 [ \n\t]* "</parameter>" root_part_1))
-root_0 ::= (([ \n\t]* "<parameter=operation>" [ \n\t]* root_prop_0 [ \n\t]* "</parameter>" root_part_0))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-root_prop_1_1 ::= ("" | ("-"))
-root_prop_1_2 ::= (([0-9] root_prop_1_2) | ([0-9]))
-root_prop_1_3 ::= ("" | ("." root_prop_1_2))
-root_prop_1_4 ::= ("" | ([+\-]))
-root_prop_1_5 ::= (([0-9] root_prop_1_5) | ([0-9]))
-root_prop_1_6 ::= ("" | ([eE] root_prop_1_4 root_prop_1_5))
-root_prop_2_1 ::= ("" | ("-"))
-root_prop_2_2 ::= (([0-9] root_prop_2_2) | ([0-9]))
-root_prop_2_3 ::= ("" | ("." root_prop_2_2))
-root_prop_2_4 ::= ("" | ([+\-]))
-root_prop_2_5 ::= (([0-9] root_prop_2_5) | ([0-9]))
-root_prop_2_6 ::= ("" | ([eE] root_prop_2_4 root_prop_2_5))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-root_prop_1_7 ::= (("0") | ([1-9] [0-9]*))
-root_prop_2_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-xml_string_1 ::= TagDispatch(
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=false,
-  excludes=("</parameter>")
-)
-xml_variable_name_1 ::= (([a-zA-Z_] [a-zA-Z0-9_]*))
-xml_string_0_1 ::= ((xml_string_1))
-xml_any_1 ::= ((basic_number_8) | (xml_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= (([ \n\t]* "<parameter=location>" [ \n\t]* xml_string_0_1 [ \n\t]* "</parameter>"))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-triggered_tags_group ::= (("Calculator>" root_0 "</function>") | ("Weather>" root_1 "</function>"))
-triggered_tags ::= TagDispatch(
-  ("<function=", triggered_tags_group),
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=true,
-  excludes=()
-)
-root ::= ((triggered_tags))
-""",
-        [
-            (
-                "OK, I will use the Calculator tool to perform the operation. <function=Calculator><parameter=operation>add</parameter><parameter=a>5</parameter><parameter=b>3</parameter></function>",
-                True,
-            ),
-            (
-                "I need to know the weather in Paris. <function=Weather><parameter=location>Paris</parameter></function>",
-                True,
-            ),
-            (
-                "Can you calculate Paris add 5? <function=Calculator><parameter=operation>add</parameter><parameter=a>5</parameter><parameter=b>Paris</parameter></function>",
-                False,
-            ),
-            ("Some random text", True),
-        ],
-    ),
-    (
-        "qwen",
-        {
-            "tools": [
-                {
-                    "name": "Calculator",
-                    "description": "A calculator that can perform basic arithmetic operations.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "operation": {
-                                "type": "string",
-                                "enum": ["add", "subtract", "multiply", "divide"],
-                            },
-                            "a": {"type": "number"},
-                            "b": {"type": "number"},
-                        },
-                        "required": ["operation", "a", "b"],
-                    },
-                },
-                {
-                    "name": "Weather",
-                    "description": "A tool to get the current weather in a specified location.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "location": {
-                                "type": "string",
-                                "description": "The name of the city or location to get the weather for.",
-                            }
-                        },
-                        "required": ["location"],
-                    },
-                },
-            ]
-        },
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_prop_0 ::= (("\"add\"") | ("\"subtract\"") | ("\"multiply\"") | ("\"divide\""))
-root_part_1 ::= (([ \n\t]* "," [ \n\t]* "\"b\"" [ \n\t]* ":" [ \n\t]* basic_number))
-root_part_0 ::= (([ \n\t]* "," [ \n\t]* "\"a\"" [ \n\t]* ":" [ \n\t]* basic_number root_part_1))
-root_0 ::= (("{" [ \n\t]* "\"operation\"" [ \n\t]* ":" [ \n\t]* root_prop_0 root_part_0 [ \n\t]* "}"))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= (("{" [ \n\t]* "\"location\"" [ \n\t]* ":" [ \n\t]* basic_string_1 [ \n\t]* "}"))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-triggered_tags_group ::= (("{\"name\": \"Calculator\", \"arguments\": " root_0 "}</tool_call>") | ("{\"name\": \"Weather\", \"arguments\": " root_1 "}</tool_call>"))
-triggered_tags ::= TagDispatch(
-  ("<tool_call>", triggered_tags_group),
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=true,
-  excludes=()
-)
-root ::= ((triggered_tags))
-""",
-        [
-            (
-                'OK, I will use the Calculator tool to perform the operation. <tool_call>{"name": "Calculator", "arguments": {"operation": "add", "a": 5, "b": 3}}</tool_call>',
-                True,
-            ),
-            (
-                'I need to know the weather in Paris. <tool_call>{"name": "Weather", "arguments": {"location": "Paris"}}</tool_call>',
-                True,
-            ),
-            (
-                'Can you calculate Paris add 5? <tool_call>{"name": "Calculator", "arguments": {"operation": "add", "a": 5, "b": "Paris"}}</tool_call>',
-                False,
-            ),
-            (
-                'I want to know the weather in 1. <tool_call>{"name": "Weather", "arguments": {"location": 1}}</tool_call>',
-                False,
-            ),
-            ("Some random text", True),
-        ],
-    ),
-    (
-        "harmony",
-        {
-            "tools": [
-                {
-                    "name": "Calculator",
-                    "description": "A calculator that can perform basic arithmetic operations.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "operation": {
-                                "type": "string",
-                                "enum": ["add", "subtract", "multiply", "divide"],
-                            },
-                            "a": {"type": "number"},
-                            "b": {"type": "number"},
-                        },
-                        "required": ["operation", "a", "b"],
-                    },
-                },
-                {
-                    "name": "Weather",
-                    "description": "A tool to get the current weather in a specified location.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "location": {
-                                "type": "string",
-                                "description": "The name of the city or location to get the weather for.",
-                            }
-                        },
-                        "required": ["location"],
-                    },
-                },
-            ],
-            "builtin_tools": [
-                {
-                    "name": "Python",
-                    "description": "A Python interpreter that can execute Python code.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "code": {"type": "string", "description": "The Python code to execute."}
-                        },
-                        "required": ["code"],
-                    },
-                }
-            ],
-        },
-        r"""any_text ::= TagDispatch(
-  stop_eos=false,
-  stop_str=("<|end|>"),
-  loop_after_dispatch=false,
-  excludes=()
-)
-any_text_1 ::= TagDispatch(
-  stop_eos=false,
-  stop_str=("<|return|>"),
-  loop_after_dispatch=false,
-  excludes=()
-)
-any_text_2 ::= TagDispatch(
-  stop_eos=false,
-  stop_str=("<|call|>"),
-  loop_after_dispatch=false,
-  excludes=()
-)
-basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_prop_0 ::= (("\"add\"") | ("\"subtract\"") | ("\"multiply\"") | ("\"divide\""))
-root_part_1 ::= (([ \n\t]* "," [ \n\t]* "\"b\"" [ \n\t]* ":" [ \n\t]* basic_number))
-root_part_0 ::= (([ \n\t]* "," [ \n\t]* "\"a\"" [ \n\t]* ":" [ \n\t]* basic_number root_part_1))
-root_0 ::= (("{" [ \n\t]* "\"operation\"" [ \n\t]* ":" [ \n\t]* root_prop_0 root_part_0 [ \n\t]* "}"))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= (("{" [ \n\t]* "\"location\"" [ \n\t]* ":" [ \n\t]* basic_string_1 [ \n\t]* "}"))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_2 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_2 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_2) | ("\\" basic_escape_2 basic_string_sub_2)) (=([ \n\t]* [,}\]:]))
-basic_any_2 ::= ((basic_number_9) | (basic_string_2) | (basic_boolean_2) | (basic_null_2) | (basic_array_3) | (basic_object_3))
-basic_integer_3 ::= (("0") | (basic_integer_1_2 [1-9] [0-9]*))
-basic_number_9 ::= ((basic_number_1_2 basic_number_7_2 basic_number_3_2 basic_number_6_2))
-basic_string_2 ::= (("\"" basic_string_sub_2))
-basic_boolean_2 ::= (("true") | ("false"))
-basic_null_2 ::= (("null"))
-basic_array_3 ::= (("[" [ \n\t]* basic_any_2 basic_array_1_2 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_3 ::= (("{" [ \n\t]* basic_string_2 [ \n\t]* ":" [ \n\t]* basic_any_2 basic_object_1_2 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_2 ::= (("{" [ \n\t]* "\"code\"" [ \n\t]* ":" [ \n\t]* basic_string_2 [ \n\t]* "}"))
-basic_integer_1_2 ::= ("" | ("-"))
-basic_number_1_2 ::= ("" | ("-"))
-basic_number_2_2 ::= (([0-9] basic_number_2_2) | ([0-9]))
-basic_number_3_2 ::= ("" | ("." basic_number_2_2))
-basic_number_4_2 ::= ("" | ([+\-]))
-basic_number_5_2 ::= (([0-9] basic_number_5_2) | ([0-9]))
-basic_number_6_2 ::= ("" | ([eE] basic_number_4_2 basic_number_5_2))
-basic_array_1_2 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_2 basic_array_1_2))
-basic_object_1_2 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_2 [ \n\t]* ":" [ \n\t]* basic_any_2 basic_object_1_2))
-basic_number_7_2 ::= (("0") | ([1-9] [0-9]*))
-triggered_tags_group ::= (("assistant<|channel|>analysis<|message|>" any_text) | ("assistant<|channel|>final<|message|>" any_text_1) | ("assistant<|channel|>final<|message|>" any_text_2) | ("assistant<|channel|>commentary to=Calculator<|constrain|>json<|message|>" root_0 "<|end|>") | ("assistant<|channel|>commentary to=Weather<|constrain|>json<|message|>" root_1 "<|end|>") | ("assistant<|channel|>analysis to=Python<|message|>" root_2 "<|end|>"))
-triggered_tags ::= TagDispatch(
-  ("<|start|>", triggered_tags_group),
-  stop_eos=true,
-  stop_str=(),
-  loop_after_dispatch=true,
-  excludes=()
-)
-root ::= ((triggered_tags))
-""",
-        [
-            (
-                '<|start|>assistant<|channel|>analysis<|message|>OK, I will use the Calculator tool to perform the operation.<|end|><|start|>assistant<|channel|>commentary to=Calculator<|constrain|>json<|message|>{"operation": "add", "a": 5, "b": 3}<|end|>',
-                True,
-            ),
-            (
-                '<|start|>assistant<|channel|>analysis<|message|>I need to know the weather in Paris.<|end|><|start|>assistant<|channel|>commentary to=Weather<|constrain|>json<|message|>{"location": "Paris"}<|end|>',
-                True,
-            ),
-            (
-                '<|start|>assistant<|channel|>analysis<|message|>Can you calculate Paris add 5?<|end|><|start|>assistant<|channel|>commentary to=Calculator<|constrain|>json<|message|>{"operation": "add", "a": 5, "b": "Paris"}<|end|>',
-                False,
-            ),
-            (
-                '<|start|>assistant<|channel|>analysis<|message|>I want to know the weather in 1.<|end|><|start|>assistant<|channel|>commentary to=Weather<|constrain|>json<|message|>{"location": 1}<|end|>',
-                False,
-            ),
-            ("<|start|>assistant<|channel|>analysis<|message|>Some random text<|end|>", True),
-            (
-                '<|start|>assistant<|channel|>analysis to=Python<|message|>{"code": "print(\\"Hello, World!\\")"}<|end|>',
-                True,
-            ),
-            (
-                "<|start|>assistant<|channel|>final<|message|>The function should be called.<|call|>",
-                True,
-            ),
-            ("<|start|>assistant<|channel|>final<|message|>All tasks done.<|return|>", True),
-        ],
-    ),
-    (
-        "llama",
-        {"tools": []},
-        r"""any_text ::= (([\0-\U0010ffff]*))
-root ::= ((any_text))
-""",
-        [
-            (
-                'OK, I will use the Calculator tool to perform the operation. {"name": "Calculator", "parameters": {"operation": "add", "a": 5, "b": 3}}',
-                True,
-            ),
-            (
-                'I need to know the weather in Paris. {"name": "Weather", "parameters": {"location": "Paris"}}',
-                True,
-            ),
-            (
-                'Can you calculate Paris add 5? {"name": "Calculator", "parameters": {"operation": "add", "a": 5, "b": "Paris"}}',
-                True,
-            ),
-            (
-                'I want to know the weather in 1. {"name": "Weather", "parameters": {"location": 1}}',
-                True,
-            ),
-            ("Some random text", True),
-        ],
-    ),
-]
-
-
-@pytest.mark.parametrize(
-    "builtin_format_type, template_values, expected_grammar, instance_is_accepted_tuples",
-    builtin_template_values_stag_grammar_instance_accepted,
-)
-def test_builtin_template_values(
-    builtin_format_type: str,
-    template_values: Dict[str, List[Dict[str, Any]]],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
-):
-    """Test builtin format with template values"""
-    template_stag_format = xgr.builtin_structural_tag_template.get_builtin_structural_tag_template(
-        builtin_format_type
-    )
-
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, **template_values)
 
 
 array_and_none_array_elements_template_value_grammar_instance_accepted = [
@@ -4060,100 +3075,35 @@ array_and_none_array_elements_template_value_grammar_instance_accepted = [
                 "functions": [{"type": "string"}, {"type": "number"}, {"type": "boolean"}],
             }
         },
-        r"""basic_escape ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub) | ("\\" basic_escape basic_string_sub)) (=([ \n\t]* [,}\]:]))
-basic_any ::= ((basic_number) | (basic_string) | (basic_boolean) | (basic_null) | (basic_array) | (basic_object))
-basic_integer ::= (("0") | (basic_integer_1 [1-9] [0-9]*))
-basic_number ::= ((basic_number_1 basic_number_7 basic_number_3 basic_number_6))
-basic_string ::= (("\"" basic_string_sub))
-basic_boolean ::= (("true") | ("false"))
-basic_null ::= (("null"))
-basic_array ::= (("[" [ \n\t]* basic_any basic_array_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object ::= (("{" [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_0 ::= ((basic_string))
-basic_integer_1 ::= ("" | ("-"))
-basic_number_1 ::= ("" | ("-"))
-basic_number_2 ::= (([0-9] basic_number_2) | ([0-9]))
-basic_number_3 ::= ("" | ("." basic_number_2))
-basic_number_4 ::= ("" | ([+\-]))
-basic_number_5 ::= (([0-9] basic_number_5) | ([0-9]))
-basic_number_6 ::= ("" | ([eE] basic_number_4 basic_number_5))
-basic_array_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any basic_array_1))
-basic_object_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string [ \n\t]* ":" [ \n\t]* basic_any basic_object_1))
-basic_number_7 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_1 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_1 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_1) | ("\\" basic_escape_1 basic_string_sub_1)) (=([ \n\t]* [,}\]:]))
-basic_any_1 ::= ((basic_number_8) | (basic_string_1) | (basic_boolean_1) | (basic_null_1) | (basic_array_2) | (basic_object_2))
-basic_integer_2 ::= (("0") | (basic_integer_1_1 [1-9] [0-9]*))
-basic_number_8 ::= ((basic_number_1_1 basic_number_7_1 basic_number_3_1 basic_number_6_1))
-basic_string_1 ::= (("\"" basic_string_sub_1))
-basic_boolean_1 ::= (("true") | ("false"))
-basic_null_1 ::= (("null"))
-basic_array_2 ::= (("[" [ \n\t]* basic_any_1 basic_array_1_1 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_2 ::= (("{" [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_1 ::= ((basic_number_8))
-basic_integer_1_1 ::= ("" | ("-"))
-basic_number_1_1 ::= ("" | ("-"))
-basic_number_2_1 ::= (([0-9] basic_number_2_1) | ([0-9]))
-basic_number_3_1 ::= ("" | ("." basic_number_2_1))
-basic_number_4_1 ::= ("" | ([+\-]))
-basic_number_5_1 ::= (([0-9] basic_number_5_1) | ([0-9]))
-basic_number_6_1 ::= ("" | ([eE] basic_number_4_1 basic_number_5_1))
-basic_array_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_1 basic_array_1_1))
-basic_object_1_1 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_1 [ \n\t]* ":" [ \n\t]* basic_any_1 basic_object_1_1))
-basic_number_7_1 ::= (("0") | ([1-9] [0-9]*))
-basic_escape_2 ::= (([\"\\/bfnrt]) | ("u" [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9] [A-Fa-f0-9]))
-basic_string_sub_2 ::= (("\"") | ([^\0-\x1f\"\\\r\n] basic_string_sub_2) | ("\\" basic_escape_2 basic_string_sub_2)) (=([ \n\t]* [,}\]:]))
-basic_any_2 ::= ((basic_number_9) | (basic_string_2) | (basic_boolean_2) | (basic_null_2) | (basic_array_3) | (basic_object_3))
-basic_integer_3 ::= (("0") | (basic_integer_1_2 [1-9] [0-9]*))
-basic_number_9 ::= ((basic_number_1_2 basic_number_7_2 basic_number_3_2 basic_number_6_2))
-basic_string_2 ::= (("\"" basic_string_sub_2))
-basic_boolean_2 ::= (("true") | ("false"))
-basic_null_2 ::= (("null"))
-basic_array_3 ::= (("[" [ \n\t]* basic_any_2 basic_array_1_2 [ \n\t]* "]") | ("[" [ \n\t]* "]"))
-basic_object_3 ::= (("{" [ \n\t]* basic_string_2 [ \n\t]* ":" [ \n\t]* basic_any_2 basic_object_1_2 [ \n\t]* "}") | ("{" [ \n\t]* "}"))
-root_2 ::= ((basic_boolean_2))
-basic_integer_1_2 ::= ("" | ("-"))
-basic_number_1_2 ::= ("" | ("-"))
-basic_number_2_2 ::= (([0-9] basic_number_2_2) | ([0-9]))
-basic_number_3_2 ::= ("" | ("." basic_number_2_2))
-basic_number_4_2 ::= ("" | ([+\-]))
-basic_number_5_2 ::= (([0-9] basic_number_5_2) | ([0-9]))
-basic_number_6_2 ::= ("" | ([eE] basic_number_4_2 basic_number_5_2))
-basic_array_1_2 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_any_2 basic_array_1_2))
-basic_object_1_2 ::= ("" | ([ \n\t]* "," [ \n\t]* basic_string_2 [ \n\t]* ":" [ \n\t]* basic_any_2 basic_object_1_2))
-basic_number_7_2 ::= (("0") | ([1-9] [0-9]*))
-or ::= ((root_0) | (root_1) | (root_2))
-tag ::= (("<calling>" or "</calling>"))
-root ::= ((tag))
-""",
-        [
-            ('<calling>"Hello, World!"</calling>', True),
-            ("<calling>42</calling>", True),
-            ("<calling>true</calling>", True),
-            ("<calling>[1, 2, 3]</calling>", False),
-            ('<calling>{"key": "value"}</calling>', False),
-        ],
+        {
+            "type": "structural_tag",
+            "format": {
+                "type": "tag",
+                "begin": "<calling>",
+                "content": {
+                    "type": "or",
+                    "elements": [
+                        {"type": "json_schema", "json_schema": {"type": "string"}},
+                        {"type": "json_schema", "json_schema": {"type": "number"}},
+                        {"type": "json_schema", "json_schema": {"type": "boolean"}},
+                    ],
+                },
+                "end": ["</calling>"],
+            },
+        },
     )
 ]
 
 
 @pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
+    "template_stag_format, template_values, expected_stag",
     array_and_none_array_elements_template_value_grammar_instance_accepted,
 )
 def test_array_and_none_array_elements(
-    template_stag_format: Dict[str, Any],
-    template_values: Dict[str, Any],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
+    template_stag_format: Dict[str, Any], template_values: Dict[str, Any], expected_stag: Dict
 ):
     """Test array and none-array elements in template values"""
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
+    check_template_stag_with_stag(template_stag_format, expected_stag, **template_values)
 
 
 def test_invalid_multiple_array():
@@ -4176,581 +3126,11 @@ def test_invalid_multiple_array():
     }
 
     with pytest.raises(Exception) as exc_info:
-        check_template_stag_with_grammar(structural_tag_format, "", **template_values)
+        check_template_stag_with_stag(structural_tag_format, "", **template_values)
     expected_error_info = (
         "Invalid structural tag error: Invalid placeholder structure found in tag format"
     )
     assert str(exc_info.value) == expected_error_info
-
-
-nested_array_template_value_grammar_instance_accepted = [
-    (
-        {
-            "type": "structural_tag",
-            "format": {
-                "type": "const_string",
-                "value": "Hello, {{nested.nested[].nested.nested[]}}!",
-            },
-        },
-        {
-            "nested": {
-                "nested": [
-                    {"nested": {"nested": ["A", "B", "C"]}},
-                    {"nested": {"nested": ["D", "E", "F"]}},
-                ]
-            }
-        },
-        r"""const_string ::= (("Hello, A!"))
-const_string_1 ::= (("Hello, B!"))
-const_string_2 ::= (("Hello, C!"))
-const_string_3 ::= (("Hello, D!"))
-const_string_4 ::= (("Hello, E!"))
-const_string_5 ::= (("Hello, F!"))
-or ::= ((const_string) | (const_string_1) | (const_string_2) | (const_string_3) | (const_string_4) | (const_string_5))
-root ::= ((or))
-""",
-        [
-            ("Hello, A!", True),
-            ("Hello, B!", True),
-            ("Hello, C!", True),
-            ("Hello, D!", True),
-            ("Hello, E!", True),
-            ("Hello, F!", True),
-            ("Hello, G!", False),
-        ],
-    ),
-    (
-        {
-            "type": "structural_tag",
-            "format": {
-                "type": "sequence",
-                "elements": [
-                    {"type": "const_string", "value": "{{nested1.nested2[].first}}"},
-                    {
-                        "type": "const_string",
-                        "value": "Hello, {{nested1.nested2[].nested3.nested4[]}}!",
-                    },
-                    {"type": "const_string", "value": "{{nested1.nested2[].nested3.third}}"},
-                    {"type": "const_string", "value": "{{nested1.last}}"},
-                ],
-            },
-        },
-        {
-            "nested1": {
-                "nested2": [
-                    {
-                        "nested3": {"nested4": ["A", "B", "C"], "third": "Farewell."},
-                        "first": "Greetings",
-                    },
-                    {
-                        "nested3": {"nested4": ["D", "E", "F"], "third": "See you later."},
-                        "first": "Salutations",
-                    },
-                ],
-                "last": "Goodbye.",
-            }
-        },
-        r"""const_string ::= (("Greetings"))
-const_string_1 ::= (("Hello, A!"))
-const_string_2 ::= (("Hello, B!"))
-const_string_3 ::= (("Hello, C!"))
-or ::= ((const_string_1) | (const_string_2) | (const_string_3))
-const_string_4 ::= (("Farewell."))
-const_string_5 ::= (("Goodbye."))
-sequence ::= ((const_string or const_string_4 const_string_5))
-const_string_6 ::= (("Salutations"))
-const_string_7 ::= (("Hello, D!"))
-const_string_8 ::= (("Hello, E!"))
-const_string_9 ::= (("Hello, F!"))
-or_1 ::= ((const_string_7) | (const_string_8) | (const_string_9))
-const_string_10 ::= (("See you later."))
-const_string_11 ::= (("Goodbye."))
-sequence_1 ::= ((const_string_6 or_1 const_string_10 const_string_11))
-or_2 ::= ((sequence) | (sequence_1))
-root ::= ((or_2))
-""",
-        [
-            ("GreetingsHello, A!Farewell.Goodbye.", True),
-            ("GreetingsHello, B!Farewell.Goodbye.", True),
-            ("GreetingsHello, C!Farewell.Goodbye.", True),
-            ("SalutationsHello, D!See you later.Goodbye.", True),
-            ("SalutationsHello, E!See you later.Goodbye.", True),
-            ("SalutationsHello, F!See you later.Goodbye.", True),
-            ("GreetingsHello, D!Farewell.Goodbye.", False),
-            ("SalutationsHello, A!See you later.Goodbye.", False),
-        ],
-    ),
-    (
-        {
-            "type": "structural_tag",
-            "format": {
-                "type": "sequence",
-                "elements": [
-                    {
-                        "type": "tag",
-                        "begin": "{{nested1.nested2[].first}}",
-                        "content": {
-                            "type": "const_string",
-                            "value": "Hello, {{nested1.nested2[].nested3.nested4[]}}!",
-                        },
-                        "end": "{{nested1.nested2[].nested3.third}}",
-                    },
-                    {"type": "const_string", "value": "{{nested1.last}}"},
-                ],
-            },
-        },
-        {
-            "nested1": {
-                "nested2": [
-                    {
-                        "nested3": {"nested4": ["A", "B", "C"], "third": "Farewell."},
-                        "first": "Greetings",
-                    },
-                    {
-                        "nested3": {"nested4": ["D", "E", "F"], "third": "See you later."},
-                        "first": "Salutations",
-                    },
-                ],
-                "last": "Goodbye.",
-            }
-        },
-        r"""const_string ::= (("Hello, A!"))
-const_string_1 ::= (("Hello, B!"))
-const_string_2 ::= (("Hello, C!"))
-or ::= ((const_string) | (const_string_1) | (const_string_2))
-tag ::= (("Greetings" or "Farewell."))
-const_string_3 ::= (("Goodbye."))
-sequence ::= ((tag const_string_3))
-const_string_4 ::= (("Hello, D!"))
-const_string_5 ::= (("Hello, E!"))
-const_string_6 ::= (("Hello, F!"))
-or_1 ::= ((const_string_4) | (const_string_5) | (const_string_6))
-tag_1 ::= (("Salutations" or_1 "See you later."))
-const_string_7 ::= (("Goodbye."))
-sequence_1 ::= ((tag_1 const_string_7))
-or_2 ::= ((sequence) | (sequence_1))
-root ::= ((or_2))
-""",
-        [
-            ("GreetingsHello, A!Farewell.Goodbye.", True),
-            ("GreetingsHello, B!Farewell.Goodbye.", True),
-            ("GreetingsHello, C!Farewell.Goodbye.", True),
-            ("SalutationsHello, D!See you later.Goodbye.", True),
-            ("SalutationsHello, E!See you later.Goodbye.", True),
-            ("SalutationsHello, F!See you later.Goodbye.", True),
-            ("GreetingsHello, D!Farewell.Goodbye.", False),
-            ("SalutationsHello, A!See you later.Goodbye.", False),
-        ],
-    ),
-]
-
-
-@pytest.mark.parametrize(
-    "template_stag_format, template_values, expected_grammar, instance_is_accepted_tuples",
-    nested_array_template_value_grammar_instance_accepted,
-)
-def test_nested_array_template(
-    template_stag_format: Dict[str, Any],
-    template_values: Dict[str, Any],
-    expected_grammar: str,
-    instance_is_accepted_tuples: List[Tuple[str, bool]],
-):
-    check_template_stag_with_grammar(template_stag_format, expected_grammar, **template_values)
-    for instance, is_accepted in instance_is_accepted_tuples:
-        check_template_stag_with_instance(
-            template_stag_format, instance, is_accepted, **template_values
-        )
-
-
-example_tools = [
-    {
-        "name": "search_web",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "query": {
-                    "type": "string",
-                    "description": "Search keywords or a full query string; must not be empty",
-                },
-                "top_k": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 50,
-                    "description": "Maximum number of results to return",
-                },
-            },
-            "required": ["query"],
-        },
-    },
-    {
-        "name": "summarize_text",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Original text to be summarized"},
-                "max_length": {
-                    "type": "integer",
-                    "minimum": 10,
-                    "description": "Maximum length of the summary in tokens or characters",
-                },
-            },
-            "required": ["text"],
-        },
-    },
-    {
-        "name": "translate_text",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Text to be translated"},
-                "target_language": {
-                    "type": "string",
-                    "description": "Target language code, e.g. en, zh, ja",
-                },
-                "source_language": {
-                    "type": "string",
-                    "description": "Source language code; optional, auto-detected if omitted",
-                },
-            },
-            "required": ["text", "target_language"],
-        },
-    },
-    {
-        "name": "extract_keywords",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Input text"},
-                "top_k": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 20,
-                    "description": "Number of keywords to extract",
-                },
-            },
-            "required": ["text"],
-        },
-    },
-    {
-        "name": "classify_text",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Text to be classified"},
-                "labels": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": "Candidate label list",
-                },
-                "multi_label": {
-                    "type": "boolean",
-                    "description": "Whether multiple labels are allowed",
-                },
-            },
-            "required": ["text", "labels"],
-        },
-    },
-    {
-        "name": "sentiment_analysis",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Text used for sentiment analysis"},
-                "granularity": {
-                    "type": "string",
-                    "enum": ["document", "sentence"],
-                    "description": "Granularity level of analysis",
-                },
-            },
-            "required": ["text"],
-        },
-    },
-    {
-        "name": "generate_code",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "prompt": {
-                    "type": "string",
-                    "description": "Description of the code generation requirement",
-                },
-                "language": {
-                    "type": "string",
-                    "description": "Target programming language, e.g. python, cpp",
-                },
-                "style": {"type": "string", "description": "Desired code style or conventions"},
-            },
-            "required": ["prompt", "language"],
-        },
-    },
-    {
-        "name": "explain_code",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "code": {"type": "string", "description": "Code snippet to be explained"},
-                "language": {"type": "string", "description": "Programming language of the code"},
-                "detail_level": {
-                    "type": "string",
-                    "enum": ["brief", "detailed"],
-                    "description": "Level of explanation detail",
-                },
-            },
-            "required": ["code"],
-        },
-    },
-    {
-        "name": "refactor_code",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "code": {"type": "string", "description": "Original code"},
-                "language": {"type": "string", "description": "Programming language"},
-                "style": {
-                    "type": "string",
-                    "description": "Refactoring goal, e.g. performance or readability",
-                },
-            },
-            "required": ["code", "language"],
-        },
-    },
-    {
-        "name": "answer_question",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "question": {"type": "string", "description": "User question"},
-                "context": {
-                    "type": "string",
-                    "description": "Optional background or reference information",
-                },
-            },
-            "required": ["question"],
-        },
-    },
-    {
-        "name": "generate_sql",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "description": {
-                    "type": "string",
-                    "description": "Natural language description of the SQL query",
-                },
-                "dialect": {"type": "string", "description": "SQL dialect, e.g. mysql, postgres"},
-                "tables": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": "Available table names",
-                },
-            },
-            "required": ["description"],
-        },
-    },
-    {
-        "name": "analyze_data",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "data": {
-                    "type": "array",
-                    "description": "Input dataset, typically a structured array",
-                },
-                "operation": {
-                    "type": "string",
-                    "description": "Analysis operation, e.g. statistics or aggregation",
-                },
-                "group_by": {"type": "string", "description": "Field used for grouping"},
-            },
-            "required": ["data", "operation"],
-        },
-    },
-    {
-        "name": "format_text",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Original text"},
-                "format": {
-                    "type": "string",
-                    "enum": ["markdown", "html", "latex"],
-                    "description": "Target output format",
-                },
-            },
-            "required": ["text", "format"],
-        },
-    },
-    {
-        "name": "detect_language",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Text whose language is to be detected"}
-            },
-            "required": ["text"],
-        },
-    },
-    {
-        "name": "generate_test_cases",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "function_description": {
-                    "type": "string",
-                    "description": "Description of function behavior and constraints",
-                },
-                "language": {"type": "string", "description": "Target language for the test cases"},
-                "count": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "description": "Number of test cases to generate",
-                },
-            },
-            "required": ["function_description", "language"],
-        },
-    },
-    {
-        "name": "rewrite_text",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Original text"},
-                "tone": {
-                    "type": "string",
-                    "description": "Desired tone, e.g. formal or conversational",
-                },
-                "length_constraint": {
-                    "type": "string",
-                    "description": "Length requirement, e.g. shorter or longer",
-                },
-            },
-            "required": ["text"],
-        },
-    },
-    {
-        "name": "extract_entities",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "text": {"type": "string", "description": "Input text"},
-                "entity_types": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": "Entity types to extract, e.g. person or location",
-                },
-            },
-            "required": ["text"],
-        },
-    },
-    {
-        "name": "generate_outline",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "topic": {"type": "string", "description": "Topic or title"},
-                "depth": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 5,
-                    "description": "Depth level of the outline",
-                },
-            },
-            "required": ["topic"],
-        },
-    },
-    {
-        "name": "chat_completion",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "messages": {
-                    "type": "array",
-                    "description": "List of conversation messages, each with role and content",
-                },
-                "temperature": {
-                    "type": "number",
-                    "minimum": 0,
-                    "maximum": 2,
-                    "description": "Controls randomness of generation",
-                },
-                "max_tokens": {
-                    "type": "integer",
-                    "description": "Maximum number of tokens to generate",
-                },
-            },
-            "required": ["messages"],
-        },
-    },
-    {
-        "name": "evaluate_answer",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "question": {"type": "string", "description": "Original question"},
-                "answer": {"type": "string", "description": "Answer to be evaluated"},
-                "criteria": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": "Evaluation criteria, e.g. correctness or completeness",
-                },
-            },
-            "required": ["question", "answer"],
-        },
-    },
-]
-
-
-builtin_model_names = ["llama", "qwen", "qwen_coder", "deepseek", "kimi"]
-function_numbers = [5, 20]
-
-
-@pytest.mark.parametrize("model_name", builtin_model_names)
-@pytest.mark.parametrize("function_number", function_numbers)
-def test_builtin_template_efficiency(model_name: str, function_number: int):
-    tools = example_tools[:function_number]
-    builtin_template = xgr.builtin_structural_tag_template.get_builtin_structural_tag_template(
-        model_name
-    )
-
-    start_time = time.monotonic_ns()
-    grammar = xgr.Grammar.from_structural_tag_template(builtin_template, tools=tools)
-    end_time = time.monotonic_ns()
-    print(
-        f"Model: {model_name}, Function number: {function_number}, Conversion Time: {(end_time - start_time) / 1e6} ms."
-    )
-
-    if profiler is not None:
-        start_time = time.monotonic_ns()
-        _ = profiler.compiler.compile_grammar(grammar)
-        end_time = time.monotonic_ns()
-        print(
-            f"Model: {model_name}, Function number: {function_number}, Compilation Time: {(end_time - start_time) / 1e6} ms."
-        )
-
-
-@pytest.mark.parametrize("function_number", function_numbers)
-def test_harmony_builtin_template_efficiency(function_number: int):
-    tools = example_tools[:function_number]
-    builtin_template = xgr.builtin_structural_tag_template.get_builtin_structural_tag_template(
-        "harmony"
-    )
-
-    start_time = time.monotonic_ns()
-    grammar = xgr.Grammar.from_structural_tag_template(
-        builtin_template, tools=tools, builtin_tools=tools
-    )
-    end_time = time.monotonic_ns()
-    print(
-        f"Model: harmony, Function number: {function_number}, Conversion Time: {(end_time - start_time) / 1e6} ms."
-    )
-
-    if profiler is not None:
-        start_time = time.monotonic_ns()
-        _ = profiler.compiler.compile_grammar(grammar)
-        end_time = time.monotonic_ns()
-        print(
-            f"Model: harmony, Function number: {function_number}, Compilation Time: {(end_time - start_time) / 1e6} ms."
-        )
 
 
 if __name__ == "__main__":
